+++ resources/js/pieChart.js
@@ -1,3 +1,4 @@
+import {mergedOptionsWithJsonConfig} from './helpers'
 
 const pieChart = () => {
     return {
@@ -15,12 +16,15 @@ const pieChart = () => {
             }
 
             const title = component.get('pieChartModel.title');
+
             const animated = component.get('pieChartModel.animated') || false;
             const dataLabels = component.get('pieChartModel.dataLabels') || {};
             const onSliceClickEventName = component.get('pieChartModel.onSliceClickEventName');
             const data = component.get('pieChartModel.data');
             const sparkline = component.get('pieChartModel.sparkline');
+            const type = component.get('pieChartModel.type');
             const config = component.get('pieChartModel.config');
+            const jsonConfig = component.get('pieChartModel.jsonConfig');
 
             const options = {
                 series: data.map(item => item.value),
@@ -28,7 +32,7 @@ const pieChart = () => {
                 chart: {
                     fontFamily: config.font_family,
                     height: '100%',
-                    type: 'pie',
+                    type: type,
 
                     ...sparkline,
 
@@ -74,6 +78,14 @@ const pieChart = () => {
                 ],
 
                 legend: component.get('pieChartModel.legend') || {},
+
+                tooltip: {
+                    y: {
+                        formatter: function(value, series) {
+                            return data[series.dataPointIndex].extras.tooltip || value;
+                        }
+                    }
+                },
             };
 
             const colors = component.get('pieChartModel.colors');
@@ -82,7 +94,7 @@ const pieChart = () => {
                 options['colors'] = colors
             }
 
-            this.chart = new ApexCharts(this.$refs.container, options);
+            this.chart = new ApexCharts(this.$refs.container, mergedOptionsWithJsonConfig(options, jsonConfig));
             this.chart.render();
         }
     }