+++ src/LivewireCharts.php
@@ -6,6 +6,8 @@ use Asantibanez\LivewireCharts\Models\AreaChartModel;
 use Asantibanez\LivewireCharts\Models\ColumnChartModel;
 use Asantibanez\LivewireCharts\Models\LineChartModel;
 use Asantibanez\LivewireCharts\Models\PieChartModel;
+use Asantibanez\LivewireCharts\Models\RadarChartModel;
+use Asantibanez\LivewireCharts\Models\TreeMapChartModel;
 
 class LivewireCharts
 {
@@ -41,5 +43,15 @@ class LivewireCharts
     public function pieChartModel()
     {
         return new PieChartModel();
+    }
+
+    public function radarChartModel()
+    {
+        return new RadarChartModel();
+    }
+
+    public function treeMapChartModel()
+    {
+        return new TreeMapChartModel();
     }
 }