+++ src/Models/BaseChartModel.php
@@ -16,6 +16,7 @@ class BaseChartModel
     use HasSparkline;
     use HasGrid;
     use HasColors;
+    use HasJsonConfig;
 
     public function __construct()
     {
@@ -28,6 +29,7 @@ class BaseChartModel
         $this->initSparkline();
         $this->initGrid();
         $this->initColors();
+        $this->initJsonConfig();
     }
 
     public function reactiveKey()
@@ -46,7 +48,8 @@ class BaseChartModel
             $this->dataLabelsToArray(),
             $this->sparklineToArray(),
             $this->gridToArray(),
-            $this->colorsToArray()
+            $this->colorsToArray(),
+            $this->jsonConfigToArray()
         );
     }
 
@@ -61,5 +64,6 @@ class BaseChartModel
         $this->sparklineFromArray($array);
         $this->gridFromArray($array);
         $this->colorsFromArray($array);
+        $this->jsonConfigFromArray($array);
     }
 }