+++ src/Models/PieChartModel.php
@@ -11,6 +11,8 @@ class PieChartModel extends BaseChartModel
 
     public $opacity;
 
+    public $type;
+
     public $data;
 
     public function __construct()
@@ -21,12 +23,21 @@ class PieChartModel extends BaseChartModel
 
         $this->opacity = 0.75;
 
+        $this->type = 'pie';
+
         $this->data = collect();
     }
 
     public function setOpacity($opacity)
     {
         $this->opacity = $opacity;
+
+        return $this;
+    }
+
+    public function setType($type)
+    {
+        $this->type = $type;
 
         return $this;
     }
@@ -55,6 +66,7 @@ class PieChartModel extends BaseChartModel
         return array_merge(parent::toArray(), [
             'onSliceClickEventName' => $this->onSliceClickEventName,
             'opacity' => $this->opacity,
+            'type' => $this->type,
             'data' => $this->data->toArray(),
             'config' => config('livewire-charts'),
         ]);
@@ -67,6 +79,8 @@ class PieChartModel extends BaseChartModel
         $this->onSliceClickEventName = data_get($array, 'onSliceClickEventName', null);
 
         $this->opacity = data_get($array, 'opacity', 0.75);
+
+        $this->type = data_get($array, 'type', 'pie');
 
         $this->data = collect(data_get($array, 'data', []));
     }